refactor(datasets): extract create/update/draft routes to routes/write.ts (6d)
createDatasetRoute/updateDatasetRoute + draft validate/cancel as typed adapters.
Adds the _draft accessor (setReqDraft/reqDraftOptional, legacyProp — upload.ts still
raw-reads req._draft). permissions.middleware trackingCategory widened to string|null
(writeData routes pass null). Two preserved-bit-for-bit suspected bugs recorded for the
parking lot: (1) cancelDraft passed a stray 5th arg to the 4-param journals.log (dropped,
no-op); (2) draft-validated localizedParams is mis-shaped ({cause:{fr,en}}).
